Update README.md
The code is a web scraper that extracts the ratings and reviews from a set of web pages at the URL base_url (which is set to "https://www.airlinequality.com/airline-reviews/british-airways"). The scraper retrieves and processes pages (which is set to 10) number of pages, each with page_size (which is set to 100) number of reviews.

The data that is collected is stored in two lists: ratings and reviews. The ratings list stores the rating values (as text strings) of the reviews, while the reviews list stores the full text content of the reviews.

The scraping process is performed in a for loop that runs pages number of times. On each iteration, the URL of the current page is constructed and the HTML content of the page is collected using the requests library. The content is then parsed using the BeautifulSoup library.

Next, the code uses the BeautifulSoup object to find all <div> elements on the page that have the attribute itemprop set to "reviewRating". For each such element, the code finds the <span> element within it that has the attribute itemprop set to "ratingValue". The text content of this <span> element is extracted and added to the ratings list.

Finally, the code uses the BeautifulSoup object again to find all <div> elements on the page that have the class "text_content", and the text content of these elements is extracted and added to the reviews list. The lengths of the ratings and reviews lists are then printed after each iteration of the for loop.